+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/resources/pack/DynamicResourcesProvider.java
@@ -88,10 +88,10 @@ public abstract class DynamicResourcesProvider implements SimplePackProvider {
 
     public void prepare() {
         boolean needsR = this.generationStrategy.needsRegeneration(packType);
-        if(needsR){
+        if (needsR) {
             this.needsRegeneration = this.packResources.clearAllResources();
-        }else{
-            this.needsRegeneration = !this.packResources.checkPathValidity();
+        } else {
+            this.needsRegeneration = !this.packResources.checkValidityAndInitialize();
         }
     }