+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/resources/pack/IEditablePackResources.java
@@ -27,7 +27,7 @@ public interface IEditablePackResources extends PackResources {
     default void commitChanges(Executor executor) {
     }
 
-    default boolean checkPathValidity() {
+    default boolean checkValidityAndInitialize() {
         return true;
     }
 }